move delete_metadata to content_delete util file
